% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyplotP.R
\name{easyplot}
\alias{easyplot}
\title{easyplot}
\usage{
easyplot(x, y)
}
\arguments{
\item{x}{A numerical variable in a dataset.}

\item{y}{A numerical variable in a dataset.}
}
\value{
A blue scatter plot using two numerical variables on the x and y axis
}
\description{
This function creates a standard scatter plot with 0.8 size blue points using ggplot. This function makes it easy to create a plot to look at the relationship between two numerical variables in a dataset.
}
\examples{
df <- data.frame(A=c(4, 8, 10, 2, 15, 12, 7, 22), B=c(6, 3, 9, 7, 6, 8, 14, 10))
easyplot(df$A, df$B)
}
